+++ examples/meta_example.rb
@@ -21,6 +21,10 @@ class MetaExample
   end
   
   def launch
+    menu('File') {
+      quit_menu_item
+    }
+    
     window('Meta-Example', 700, 500) { |w|
       margined true