+++ examples/color_button.rb
@@ -4,18 +4,11 @@ require 'glimmer-dsl-libui'
 
 include Glimmer
 
-window('color_button') {
-  vertical_box {
-  c = color_button { |cb|
+window('color button', 200) {
+  color_button { |cb|
     on_changed do
       rgba = cb.color
       p rgba
     end
-  }
-  button('destroy') {
-    on_clicked do
-      c.destroy
-    end
-  }
   }
 }.show