+++ examples/font_button.rb
@@ -4,16 +4,8 @@ require 'glimmer-dsl-libui'
 
 include Glimmer
 
-menu('File') {
-  menu_item('Destroy') {
-    on_clicked do
-      @f.destroy
-    end
-  }
-}
-
 window('hello world', 300, 200) {
-  @f = font_button { |fb|
+  font_button { |fb|
     on_changed do
       font_descriptor = fb.font
       p font_descriptor